# Minor optimization and cleanup for extension methods