--- a/osu.Framework/Extensions/ExtensionMethods.cs
+++ b/osu.Framework/Extensions/ExtensionMethods.cs
@@ -235,6 +235,17 @@ namespace osu.Framework.Extensions
                     .GetCustomAttribute<DescriptionAttribute>()?.Description
                ?? value.ToString();
 
+        private static string toLowercaseHex(this byte[] bytes)
+        {
+            // Convert.ToHexString is upper-case, so we are doing this ourselves
+
+            return string.Create(bytes.Length * 2, bytes, (span, b) =>
+            {
+                for (int i = 0; i < b.Length; i++)
+                    _ = b[i].TryFormat(span[(i * 2)..], out _, "x2");
+            });
+        }
+
         /// <summary>
         /// Gets a SHA-2 (256bit) hash for the given stream, seeking the stream before and after.
         /// </summary>
@@ -247,7 +258,7 @@ namespace osu.Framework.Extensions
             stream.Seek(0, SeekOrigin.Begin);
 
             using (var alg = SHA256.Create())
-                hash = BitConverter.ToString(alg.ComputeHash(stream)).Replace("-", "").ToLowerInvariant();
+                hash = alg.ComputeHash(stream).toLowercaseHex();
 
             stream.Seek(0, SeekOrigin.Begin);
 
@@ -262,7 +273,7 @@ namespace osu.Framework.Extensions
         public static string ComputeSHA2Hash(this string str)
         {
             using (var alg = SHA256.Create())
-                return BitConverter.ToString(alg.ComputeHash(new UTF8Encoding().GetBytes(str))).Replace("-", "").ToLowerInvariant();
+                return alg.ComputeHash(Encoding.UTF8.GetBytes(str)).toLowercaseHex();
         }
 
         public static string ComputeMD5Hash(this Stream stream)
@@ -271,7 +282,7 @@ namespace osu.Framework.Extensions
 
             stream.Seek(0, SeekOrigin.Begin);
             using (var md5 = MD5.Create())
-                hash = BitConverter.ToString(md5.ComputeHash(stream)).Replace("-", "").ToLowerInvariant();
+                hash = md5.ComputeHash(stream).toLowercaseHex();
             stream.Seek(0, SeekOrigin.Begin);
 
             return hash;
@@ -279,17 +290,8 @@ namespace osu.Framework.Extensions
 
         public static string ComputeMD5Hash(this string input)
         {
-            StringBuilder hash = new StringBuilder();
-
             using (var md5 = MD5.Create())
-            {
-                byte[] bytes = md5.ComputeHash(new UTF8Encoding().GetBytes(input));
-
-                for (int i = 0; i < bytes.Length; i++)
-                    hash.Append(bytes[i].ToString("x2"));
-
-                return hash.ToString();
-            }
+                return md5.ComputeHash(Encoding.UTF8.GetBytes(input)).toLowercaseHex();
         }
 
         public static DisplayIndex GetIndex(this DisplayDevice display)