nit: I know how to maximize performance, but it doesn't worth for code complexity.