@peppy 2021-07-15
Generally we ask for a simple benchmark comparison when pure optimisations are applied. Can you add one to the benchmarks project and show a before-after comparison?