@huoyaoyuan 2021-07-15
Benchmark added.

master:

|    Method |       Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |-----------:|---------:|---------:|-------:|------:|------:|----------:|
| StringMD5 |   900.8 ns |  5.64 ns |  5.28 ns | 0.1392 |     - |     - |   1,168 B |
| StringSHA | 1,200.7 ns | 11.01 ns | 10.30 ns | 0.1087 |     - |     - |     912 B |
| StreamMD5 | 2,435.1 ns | 20.10 ns | 17.82 ns | 0.0572 |     - |     - |     504 B |
| StreamSHA | 5,045.0 ns | 25.58 ns | 23.93 ns | 0.0839 |     - |     - |     760 B |

PR:

|    Method |       Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |-----------:|---------:|---------:|-------:|------:|------:|----------:|
| StringMD5 |   749.5 ns | 14.21 ns | 13.96 ns | 0.0420 |     - |     - |     352 B |
| StringSHA | 1,180.3 ns |  9.00 ns |  8.42 ns | 0.0534 |     - |     - |     448 B |
| StreamMD5 | 2,365.3 ns |  7.56 ns |  7.07 ns | 0.0343 |     - |     - |     296 B |
| StreamSHA | 5,052.7 ns | 24.95 ns | 22.12 ns | 0.0458 |     - |     - |     392 B |